+++ src/main/java/com/zy/dao/AdministratorDao.java
@@ -9,7 +9,7 @@ import org.apache.ibatis.annotations.Update;
 @Mapper
 public interface AdministratorDao {
     @Select("SELECT * FROM administrator WHERE id=#{id}")
-    public Administrator getAdministratorById(Integer id);
+    public Administrator getAdministratorById(Long id);
 
     @Update("update administrator set nick_name=#{nick_name} where id=#{id}")
     public Integer updateAdministratorInfo(Administrator administrator);
