+++ src/main/java/com/zy/domain/Essay.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class Essay {
     private Integer id;
-    private Integer publisher_id;
+    private Long publisher_id;
     private String publisher_name;
     private Long create_time;
     private String essay_name;
