+++ src/main/java/com/zy/domain/EssayMessage.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class EssayMessage {
     private Integer id;
-    private Integer messenger_id;
+    private Long messenger_id;
     private String messenger_name;
     private Long create_time;
